Youtube-Djikstras-Sagemath
This was one of my first attempts to write Djikstras algorithm.  I used Sagemath cloud which automates a lot of the graph bits such as getting the neighbors of the vertices which have already been seen.  I believe the examples come from Sarada Herke's youtube channel.  The link is below.

https://www.youtube.com/user/DrSaradaHerke