Coursera Algorithms Part 1 - twosum
In this exercise we were supposed to find two numbers that summed up to a number that was less than abs(10,000).  The algorithm as explained in the course was a polynomial that ran in O(n^2) time, but took too long on the provided dataset.  My solution was to game the dataset and only look for possibilities that might be answers.  This could only be done because the numbers are uniformly distributed around 0 (or close enough that this works).