Coursera Algorithms Part 2 - Knapsack
Here I am solving the knapsack problem.  My solution for the big knapsack basically fills in any points that don't need to be checked.  While this works in this case, if one had many small items this algorithm would get bogged down and have to check a lot of points.  But I was able to solve the problem in a reasonable amount of time on a cloud server (Sagemath Cloud).