Kaggle - Allstate Competition
I did a quick principal component analysis on the Kaggle Allstate competition.  This analysis doesn't really do anything to help one solve the problem, but I wanted to run PCA on a real dataset just for kicks.  I'm not going to include the dataset, but you can get it in the link below.

https://www.kaggle.com/c/allstate-claims-severity